Carousel, updated FWF project section, new pictures for FWF
changed the carousel to not zoom in so much when hovering

added in additional comments under the FWF photos page to explain more and link team members github.

New pictures for FWF

added instagram handle button below FWF as well

Updated my resume to say something in a better way.